requirements.txt
Django==3.2.8
django-admin-interface==0.17.1
django-ckeditor==5.9.0
django-countries==7.2.1
django-crispy-forms==1.13.0
mysqlclient==2.0.3
Pillow==8.3.2